requirements.txt
annotated-types==0.7.0
anyio==4.5.2
boto3==1.37.31
botocore==1.37.31
certifi==2025.1.31
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-sessions==0.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
MarkupSafe==2.1.5
openai==1.72.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
tqdm==4.67.1
typing-extensions==4.13.1
urllib3==1.26.20
uvicorn==0.33.0
